add RestController

+++ src/main/java/com/example/demospringboot/repository/MessageRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface MessageRepo extends JpaRepository<Message, Long> {
+public interface MessageRepo extends JpaRepository<Message, Integer> {
 
     List<Message> findByTagContainingOrTextContaining(String tag, String text);
